refactor extract registering listeners into modules
